Return an object from dispatch()
Resolves #34 and #35.

The previous dispatch() returned a simple tuple containing a json-rpc response
and an http status. An object gives the user more detailed information from the
processing of the request.

This change required a new response module. The exceptions had to be rewritten.
Finally the dispatch() function was updated to return the Response object.