Updated method of logging exception
Exceptions raised while calling a method in Request.process were being logged using just the logging utiliy, which relied on a logging handler being defined. In the event a logging handler is not defined and an exception is raised while calling a method, an error about the lack of a logging handler is raised, and the original exception is never logged.

This change makes use the _log utility, which takes into account situations where a logging handler is not defined. In the scenario described before, a stacktrace of the initial error is logged as expected.