Remove legacy exception handlers
KeyError and TypeError are very generic and might point to an error in
the method implementation rather than in the client. The new exceptions
MethodNotFoundError and InvalidArgumentsError avoid these issues, so
there is no need to catch the generic ones anymore. This effectively
changes method behavior if methods raise KeyErrors and/or TypeErrors.