Improve docs landing page visual structure
The two most important pieces of information in the docs (links to
examples and guide) were the least visible - at the end of longest
paragraph of text on the page just after shiny code examples.

I've introduced h2 headings for the important links while preserving the
original flow of the document.

This should improve docs navigation slightly.